feat(jit): implement register caching for spilled eBPF stack slots

+++ firmware/include/utils.h
@@ -2,6 +2,7 @@
 #define UTILS_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 void uart_print(const char *s);
 void uart_print_int(int n);
@@ -9,7 +10,7 @@ void uart_print_uint64(uint64_t n);
 void uart_print_hex(uint32_t n);
 void uart_print_char(char c);
 
-void *memcpy(void *dest, const void *src, uint32_t n);
-void *memset(void *s, int c, uint32_t n);
+void *memcpy(void *dest, const void *src, size_t n);
+void *memset(void *s, int c, size_t n);
 
 #endif // UTILS_H
