+++ firmware/src/utils.c
@@ -1,5 +1,6 @@
 #include "utils.h"
 #include <stdint.h>
+#include <stddef.h>
 
 // Virtual Console in Shared RAM (visible via SSH through host_manager)
 #define VCON_BASE  0x222400000ULL
@@ -45,14 +46,14 @@ void uart_print_hex(uint32_t n) {
     for (int i = 7; i >= 0; i--) uart_print_char(hex[(n >> (i * 4)) & 0xF]);
 }
 
-void *memcpy(void *dest, const void *src, uint32_t n) {
+void *memcpy(void *dest, const void *src, size_t n) {
     uint8_t *d = (uint8_t *)dest;
     const uint8_t *s = (const uint8_t *)src;
     while (n--) *d++ = *s++;
     return dest;
 }
 
-void *memset(void *s, int c, uint32_t n) {
+void *memset(void *s, int c, size_t n) {
     uint8_t *p = (uint8_t *)s;
     while (n--) *p++ = (uint8_t)c;
     return s;
